feat: SCAN-3007, refactor ContractContent and ContractAbi Component
