+++ src/app/containers/AddressContractDetail/ContractContent.tsx
@@ -1,11 +1,9 @@
-import React, { useState, useEffect } from 'react';
+import React, { useState } from 'react';
 import { trackEvent } from 'utils/ga';
 import { ScanEvent } from 'utils/gaConstants';
 import { useTranslation } from 'react-i18next';
 import { translations } from 'locales/i18n';
-import { CFX } from 'utils/constants';
 import { ContractAbi } from 'app/components/ContractAbi/Loadable';
-import { formatType } from 'js-conflux-sdk/src/contract/abi';
 import styled from 'styled-components/macro';
 import AceEditor from 'react-ace';
 import 'ace-builds/webpack-resolver';
@@ -202,106 +200,7 @@ const StyledContractContentCodeWrapper = styled.div`
 export const ContractContent = ({ contractInfo }) => {
   const { t } = useTranslation();
   const { abi, address, verify = {} } = contractInfo;
-  const [dataForRead, setDataForRead] = useState([]);
-  const [dataForWrite, setDataForWrite] = useState([]);
   const [activeKey, setActiveKey] = useState('code');
-
-  let abiJson = [];
-  try {
-    abiJson = JSON.parse(abi);
-  } catch (error) {}
-  const clickHandler = (btnType: React.SetStateAction<string>) => {
-    setActiveKey(btnType);
-    if (btnType) {
-      trackEvent({
-        category: ScanEvent.tab.category,
-        action:
-          ScanEvent.tab.action[
-            `contract${
-              (btnType + '').charAt(0).toUpperCase() + (btnType + '').slice(1)
-            }`
-          ],
-      });
-    }
-  };
-
-  // init contract object by abi and address
-  const contract = CFX.Contract({
-    abi: abiJson,
-    address,
-  });
-  useEffect(() => {
-    getReadWriteData(abiJson).then(res => {
-      const [dataForR, dataForW] = res;
-      setDataForRead(dataForR as any);
-      setDataForWrite(dataForW as any);
-    });
-    // eslint-disable-next-line react-hooks/exhaustive-deps
-  }, [contractInfo]);
-
-  async function getReadWriteData(abiJson) {
-    let dataForRead: object[] = [];
-    let dataForWrite: object[] = [];
-    let proArr: object[] = [];
-    if (Array.isArray(abiJson)) {
-      for (let abiItem of abiJson) {
-        if (abiItem.name !== '' && abiItem.type === 'function') {
-          const stateMutability = abiItem.stateMutability;
-          switch (stateMutability) {
-            case 'pure':
-            case 'view':
-              if (abiItem.inputs && abiItem.inputs.length === 0) {
-                const fullNameWithType = formatType({
-                  name: abiItem['name'],
-                  inputs: abiItem['inputs'],
-                });
-                proArr.push(contract[fullNameWithType]());
-              }
-              dataForRead.push(abiItem);
-              break;
-            case 'nonpayable':
-              dataForWrite.push(abiItem);
-              break;
-            case 'payable':
-              const payableObjs = [
-                {
-                  internalType: 'cfx',
-                  name: abiItem['name'],
-                  type: 'cfx',
-                },
-              ];
-              abiItem['inputs'] = payableObjs.concat(abiItem['inputs']);
-              dataForWrite.push(abiItem);
-              break;
-            default:
-              break;
-          }
-        }
-      }
-      const list = await Promise.allSettled(proArr);
-      let i = 0;
-      dataForRead.forEach(function (dValue, dIndex) {
-        if (dValue['inputs'].length === 0) {
-          const listItem = list[i];
-          const status = listItem['status'];
-          if (status === 'fulfilled') {
-            const val = listItem['value'];
-            if (dValue['outputs'].length > 1) {
-              dValue['value'] = val;
-            } else {
-              const arr: any = [];
-              arr.push(val);
-              dValue['value'] = arr;
-            }
-          } else {
-            dValue['error'] = listItem['reason']['message'];
-          }
-          ++i;
-        }
-      });
-    }
-    return [dataForRead, dataForWrite];
-  }
 
   let tabs = [
     {
@@ -323,6 +222,21 @@ export const ContractContent = ({ contractInfo }) => {
     ]);
   }
 
+  const clickHandler = (btnType: React.SetStateAction<string>) => {
+    setActiveKey(btnType);
+    if (btnType) {
+      trackEvent({
+        category: ScanEvent.tab.category,
+        action:
+          ScanEvent.tab.action[
+            `contract${
+              (btnType + '').charAt(0).toUpperCase() + (btnType + '').slice(1)
+            }`
+          ],
+      });
+    }
+  };
+
   return (
     <>
       <ContractBody>
@@ -335,20 +249,10 @@ export const ContractContent = ({ contractInfo }) => {
         <ContractCard>
           {activeKey === 'code' && <Code contractInfo={contractInfo} />}
           {activeKey === 'read' && (
-            <ContractAbi
-              type="read"
-              data={dataForRead}
-              contractAddress={address}
-              contract={contract}
-            ></ContractAbi>
+            <ContractAbi type="read" address={address} abi={abi}></ContractAbi>
           )}
           {activeKey === 'write' && (
-            <ContractAbi
-              type="write"
-              data={dataForWrite}
-              contractAddress={address}
-              contract={contract}
-            ></ContractAbi>
+            <ContractAbi type="write" address={address} abi={abi}></ContractAbi>
           )}
         </ContractCard>
       </ContractBody>
